PLUT-416 Prepared some files to run prettier and did other tasks
- Got rid of some compiler warnings
about some local variables having the same names as other local
variables and so the outermost local variables were being overshadowed
by the inner local variables.

- Updated npm packages

- added a .prettierignore that is just a
soft link to .gitignore , in accordance with
https://stackoverflow.com/a/73687231/1838240

- deleted an html error on poker.html
that was preventing prettier from running on it.
Still haven't run prettier.

- added `/* HTML */ ` to the HTML strings
in poker.js, so that prettier can format them like HTML. Still need to
run prettier. Will do that in next commit

- Set prettier setting `endOfLine: "auto"`.
This will make development easier on Windows, with the way that we
Windows-using developers are doing things.