PLUT-471 Set up prettier, according to some of David Ludlow's preferences
Also, I edited IntelliJ format settings by following
https://www.jetbrains.com/help/idea/prettier.html#ws_prettier_apply_code_style
(running the action on .prettierrc.json). As it says on the linked webpage, This
is just to make it so code that IntelliJ generates is probably already formatted.

Also, I added `.idea/kotlin.xml`. IntelliJ keeps automatically
generating this file, so we may as well add it to source control